+++ src/main/java/net/spell_engine/fabric/client/FabricClientMod.java
@@ -98,8 +98,11 @@ public class FabricClientMod implements ClientModInitializer {
 
         ParticleFactoryRegistry.getInstance().register(SpellEngineParticles.smoke_large.particleType(), SpellSmokeParticle.CosySmokeFactory::new);
 
-        ParticleFactoryRegistry.getInstance().register(SpellEngineParticles.swirl.particleType(), SpellAreaParticle.Factory::new);
-//        ParticleFactoryRegistry.getInstance().register(SpellEngineParticles.swirl.particleType(), SpellExplosionParticle.TemplateFactory::new);
+        for (var entry: SpellEngineParticles.areaEffects()) {
+            ParticleFactoryRegistry.getInstance().register(
+                    entry.particleType(), (provider) -> new SpellAreaParticle.Factory(provider, entry.texture())
+            );
+        }
 
         ModelLoadingPlugin.register(pluginCtx -> {
             pluginCtx.addModels(CustomModelRegistry.modelIds);