+++ src/main/java/net/spell_engine/fx/SpellEngineParticles.java
@@ -143,6 +143,16 @@ public class SpellEngineParticles {
         return variants;
     });
 
+    private static final ArrayList<TemplateEntry> area_effects = new ArrayList<>();
+    public static List<TemplateEntry> areaEffects() {
+        return area_effects;
+    }
+    public static TemplateEntry addAreaEffect(TemplateEntry entry) {
+        addTemplate(entry);
+        area_effects.add(entry);
+        return entry;
+    }
+
     /**
      * WARNING! This method is very slow, only to be used for data file generation!
      */
@@ -173,7 +183,7 @@ public class SpellEngineParticles {
     public static final TemplateEntry smoke_medium = addTemplate(new TemplateEntry("smoke_medium", Texture.of("smoke_medium", 9)));
     public static final TemplateEntry smoke_large = addTemplate(new TemplateEntry("smoke_large", Texture.vanilla("big_smoke", 12)));
 
-    public static final TemplateEntry swirl = addTemplate(new TemplateEntry("swirl", Texture.of("swirl", 16)));
+    public static final TemplateEntry swirl = addAreaEffect(new TemplateEntry("swirl", Texture.of("swirl", 16)));
 
     @Deprecated
     public static final Entry weakness_smoke = add(new Entry("weakness_smoke", Texture.of("smoke_medium", 9)));